pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labmim-micrometeorology"
version = "1.3.1"
description = "LabMiM/UFBA micrometeorology data processing, WRF output analysis, solar radiation bias correction, and ML-based model evaluation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.14"
authors = [
    {name = "Bruno Mascarenhas", email = "brunomasck@gmail.com"},
    {name = "LabMiM - UFBA"},
]
keywords = ["micrometeorology", "WRF", "netcdf", "cartopy", "geospatial", "solar-radiation", "bias-correction"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=2.0",
    "pandas>=3.0",
    "xarray>=2024.6",
    "dask[array]>=2026.0",
    "netCDF4>=1.7.4",
    "matplotlib>=3.9",
    "cartopy>=0.23",
    "seaborn>=0.13.2",
    "scikit-learn>=1.5",
    "scipy>=1.13",
    "click>=8.1",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "imageio>=2.34",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "pandas-stubs",
    "pytest-xdist>=3.5",
]
video = [
    "moviepy>=2.2",
]
# TCC: ML-based solar radiation correction (CPU-only PyTorch)
tcc = [
    "torch>=2.6.0",
    "joblib>=1.3",
    "tqdm>=4.66",
    "tensorboard>=2.14",
]
# TCC with CUDA support (install separately: pip install torch --index-url https://download.pytorch.org/whl/cu121)
tcc-cuda = [
    "joblib>=1.3",
    "tqdm>=4.66",
]

[project.scripts]
# Micrometeorology CLIs
labmim-wrf-figures = "micrometeorology.cli.render_wrf_maps:main"
labmim-wrf-geojson = "micrometeorology.cli.export_wrf_geojson:main"
labmim-sensor-process = "micrometeorology.cli.ingest_sensor_data:main"
labmim-comparison = "micrometeorology.cli.compare_wrf_observations:main"
labmim-metrics = "micrometeorology.cli.compute_metrics:main"
labmim-station-graphs = "micrometeorology.cli.generate_station_graphs:main"
# TCC CLIs
solrad-run = "solrad_correction.cli:run_experiment_cli"
solrad-colab = "solrad_correction.cli_colab:run_colab_cli"

[tool.hatch.build.targets.wheel]
packages = ["src/micrometeorology", "src/solrad_correction"]

[tool.ruff]
src = ["src"]
target-version = "py314"
line-length = 100
exclude = [
    "legacy",
    ".venv",
    "venv",
    "env",
    ".git"
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PERF", # perflint — catch performance anti-patterns
    "RUF",  # ruff-specific rules
    "C4",   # flake8-comprehensions
    "ARG",  # flake8-unused-arguments
    "PT",   # flake8-pytest-style
    "PD",   # pandas-vet
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["micrometeorology", "solrad_correction"]

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = false
strict_equality = true
extra_checks = true
ignore_missing_imports = true
exclude = ["^legacy/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
